+++ views/dataset/pages/add_identifier.templ
@@ -4,9 +4,9 @@ import (
 	"github.com/ugent-library/biblio-backoffice/ctx"
 	"github.com/ugent-library/biblio-backoffice/models"
 	"github.com/ugent-library/biblio-backoffice/views"
+	"github.com/ugent-library/biblio-backoffice/views/aria"
 	datasetviews "github.com/ugent-library/biblio-backoffice/views/dataset"
 	datasetsummaryviews "github.com/ugent-library/biblio-backoffice/views/dataset/summary"
-	"github.com/ugent-library/biblio-backoffice/views/form/aria"
 )
 
 type AddIdentifierArgs struct {
