+++ views/form/form.templ
@@ -3,7 +3,7 @@ package form
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/ugent-library/biblio-backoffice/views/form/aria"
+	"github.com/ugent-library/biblio-backoffice/views/aria"
 )
 
 type Theme int
