+++ views/publication/add_project.templ
@@ -3,7 +3,7 @@ package publication
 import (
 	"github.com/ugent-library/biblio-backoffice/ctx"
 	"github.com/ugent-library/biblio-backoffice/models"
-	"github.com/ugent-library/biblio-backoffice/views/form/aria"
+	"github.com/ugent-library/biblio-backoffice/views/aria"
 )
 
 templ AddProject(c *ctx.Ctx, publication *models.Publication, hits []*models.Project) {
