+++ views/search.templ
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"github.com/ugent-library/biblio-backoffice/ctx"
 	"github.com/ugent-library/biblio-backoffice/models"
+	"github.com/ugent-library/biblio-backoffice/views/aria"
 	"github.com/ugent-library/biblio-backoffice/views/form"
-	"github.com/ugent-library/biblio-backoffice/views/form/aria"
 	"github.com/ugent-library/biblio-backoffice/vocabularies"
 	"io"
 )
